refactor(graph_slice): extract value-level parsing helpers into parse_helpers module
Moved all value-level command-line argument parsing helper functions and the RawValue struct
from parse.rs to a new parse_helpers.rs module. This improves modularity by separating
low-level parsing logic from higher-level request building, clarifies code organization,
and maintains existing functionality with more focused responsibility in each module.
Updated parse.rs to import the helpers from the new module and added parse_helpers mod declaration.
